Added FoE_SxBuildings table
Based on the description of the CxBuildings table...
This is the first version of the supply-generating buildings. At this
point it is not yet decided if Special or Great Buildings should be
added here, in their own table, or both.

The table also includes calculated fields - currently unverified just
how accurate or useful they are.

After running the new script and creating the FoE_SxBuildings table
sqlite responds with the message "Rows affected: 65"
After running the ENTIRE script the response is "Rows affected: 157"